Termiando Logica 7 y medio

+++ app/src/main/java/com/example/sieteymedio/CardGames/data/Carta.kt
@@ -4,6 +4,6 @@ import androidx.annotation.DrawableRes
 data class Carta(
     var palo: Palos,
     var nombre:Naipes,
-    var puntos:Float,
+    var puntos: Double,
     @DrawableRes var id:Int
 )
